Testing/requirements.txt
paramiko
pytest
pyserial>=3.5
python-can>=3.3.4
dill>=0.3.4
setuptools>=57.0.0
scp
dotenv